added new rules
IMPORTANT: the rule matching logic is not relying any
longer on the relevant files extracted from the
feature_extractor (it is rather considering the data of the
underlying raw commit object).

This is one more case that motivates a future refactoring of
feature extraction and rules (the former may be absorbed in
the latter, and the future ML module could use rule results as
features).